requirements.txt
GDAL==3.2.2
imutils==0.5.4
matplotlib==3.3.2
mmcv==2.0.0
natsort==8.3.1
numpy==1.19.5
numpy==1.20.3
opencv_python==4.4.0.44
Pillow==8.0.1
Pillow==10.1.0
scikit_image==0.17.2
scikit_learn==0.24.2
scipy==1.5.3
timm==0.9.2
torch==1.13.1
torchvision==0.14.1
